+++ onlineshop/src/App.js
@@ -1,0 +1,15 @@
+import logo from "./logo.svg";
+import NavBar from "./components/NavBar";
+import "./App.css";
+import HomePage from "./components/HomePage";
+
+function App() {
+  return (
+    <div className="App">
+      <NavBar />
+      <HomePage />
+    </div>
+  );
+}
+
+export default App;
