--- src/App.js
@@ -1,11 +1,0 @@
-import "./App.css";
-
-function App() {
-  return (
-    <div className="App">
-      Hello
-    </div>
-  );
-}
-
-export default App;
